refactor(web): split timeline responsibilities
